# kardo-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KARDO API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/comments/videos/{videoId}": {
      "get": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Получение списка комментариев к видео-клипу.",
        "operationId": "getCommentsByVideoId",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Количество элементов, которые нужно пропустить для формирования текущего набора",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество элементов в наборе",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Список комментариев пользователя получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Добавление нового комментария к видео-клипу.",
        "operationId": "createComment",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Комментарий добавлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/videos/{videoId}": {
      "get": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Получение видео-клипа.",
        "operationId": "getVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoClipDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Удаление видео-клипа.",
        "operationId": "deleteVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип удален."
          }
        }
      },
      "patch": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Обновление видео-клипа.",
        "operationId": "updateVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoClipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип обновлён.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoClipDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/{commentId}": {
      "get": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Получение комментария к видео-клипу по ИД.",
        "operationId": "getCommentById",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "id комментария",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Комментарий не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Комментарий найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Удаление комментария к видео-клипу.",
        "operationId": "deleteCommentById",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "id комментария",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Комментарий не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Комментарий удалён."
          }
        }
      },
      "patch": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Обновление пользователем своего комментария к видео-клипу по ИД.",
        "operationId": "updateCommentById",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "id комментария",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Комментарий не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Комментарий обновлён.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "type": "string"
          },
          "avatarPhoto": {
            "type": "string"
          }
        },
        "description": "Сущность пользователя короткая."
      },
      "VideoClipDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "publicationTime": {
            "type": "string",
            "format": "date-time"
          },
          "hashtags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "creatorDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "likesCount": {
            "type": "integer",
            "format": "int32"
          },
          "videoLink": {
            "type": "string"
          }
        },
        "description": "Сущность видео-клипа."
      },
...
      "NewCommentRequest": {
        "required": [
          "authorId",
          "text"
        ],
        "type": "object",
        "properties": {
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "text": {
            "maxLength": 500,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные добавляемого комментария"
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "authorDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "publicationTime": {
            "type": "string",
            "format": "date-time"
          },
          "videoClipDto": {
            "$ref": "#/components/schemas/VideoClipDto"
          },
          "text": {
            "type": "string"
          }
        },
        "description": "Сущность комментария к видео-клипу."
      },
...
      "UpdateVideoClipRequest": {
        "required": [
          "hashtags"
        ],
        "type": "object",
        "properties": {
          "hashtags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Данные обновляемого видео-клипа"
      },
...
      "UpdateCommentRequest": {
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 500,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные обновляемого комментария"
      },
...
    }
  }
}